Add SignalR, Start some Styling, update Gateways

+++ EenJaarGratis.Service.Handlers/Handlers/Question/DeleteQuestionGroupHandler.cs
@@ -8,7 +8,7 @@ using MediatR;
 
 namespace EenJaarGratis.Services.Handlers.Handlers.Question;
 
-public class DeleteQuestionGroupHandler : IRequestHandler<DeleteQuestionGroupRequest, QuestionGroupResponse>
+public class DeleteQuestionGroupHandler : IRequestHandler<DeleteQuestionGroupRequest>
 {
     private readonly IMapper _mapper;
     private readonly IQuestionGroupRepository _questionGroupRepository;
@@ -19,10 +19,14 @@ public class DeleteQuestionGroupHandler : IRequestHandler<DeleteQuestionGroupReq
         _questionGroupRepository = questionGroupRepository;
     }
 
-    public async Task<QuestionGroupResponse> Handle(DeleteQuestionGroupRequest request, CancellationToken cancellationToken)
+    public async Task<Unit> Handle(DeleteQuestionGroupRequest request, CancellationToken cancellationToken)
     {
         QuestionGroup questionGroup = await _questionGroupRepository.GetById(request.QuestionGroupId);
+        if (questionGroup is null)
+        {
+            throw new KeyNotFoundException($"Groep met id {request.QuestionGroupId} niet gevonden");
+        }
         await _questionGroupRepository.Delete(questionGroup, cancellationToken);
-        return _mapper.Map<QuestionGroupResponse>(questionGroup);
+        return Unit.Value;
     }
 }
